requirements.txt
numpy==1.14.3
opencv-python==3.4.0.12
pandas==0.22.0
Pillow==5.2.0
pretrainedmodels==0.6.2
python-telegram-bot==10.1.0
sacred==0.7.4
scikit_image==0.13.1
scikit_learn==0.19.2
scipy==1.1.0
torch==0.4.0
torchvision==0.2.1
tqdm==4.23.3
git+https://github.com/aleju/imgaug
